feat: implement run_dim_reduction script for executing dimensionality reduction jobs with YAML configuration
